LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-generated-content-expected.txt

A
Y A
XYX

FAIL fieldset generated content assert_equals: clientWidth expected 42 but got 28
